requirements.txt
fastapi==0.116.1
mysql-connector-python==9.4.0
numpy==2.2.6
pandas==2.2.3
pydantic==2.11.7
pytest==8.4.2
requests==2.32.3
seaborn==0.13.2
streamlit==1.50.0
uvicorn==0.35.0
